+++ resources/views/email/newRequest.blade.php
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+
+<html>
+
+<body>
+
+    <h1>Здравствуйте, уважаемый студент!</h1>
+
+    <p>Вы подали заявку на сайте https://hatalmasu.almau.edu.kz</p>
+
+    <p>Ваш номер заявки: <b>{{ $supportTicketID }}</b></p>
+
+    <p>Вы можете проверить статус заявки на странице https://hatalmasu.almau.edu.kz/SupportTicketStatus
+</body>
+
+</html>