+++ resources/views/email/statusUpdate.blade.php
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+
+<html>
+
+<body>
+
+    <h1>Здравствуйте, уважаемый студент!</h1>
+
+    <p>Ваша заявка <b>{{ $supportTicketStatus }}</b></p>
+
+    <p>{{ $comment }}</p>
+
+    <p>Ваш номер заявки: <b>{{ $supportTicketID }}</b></p>
+
+    <p>Вы можете проверить статус заявки на странице https://hatalmasu.almau.edu.kz/SupportTicketStatus
+</body>
+
+</html>