+++ routes/web.php
@@ -8,7 +8,6 @@ use Illuminate\Support\Facades\Session;
 use TCG\Voyager\Facades\Voyager;
 
 use App\Http\Controllers\SupportTicketController;
-use App\Http\Controllers\DormitoryTicket;
 
 /*
 |--------------------------------------------------------------------------
@@ -31,7 +30,6 @@ Route::view('/Info', 'Info');
 Route::view('/info', 'Info');
 
 Route::group(['prefix' => 'admin'], function () {
-
     //Choose tickets by status (should be before voyager routes)
     Route::get('/supporttickets_{status}', [SupportTicketController::class, 'chooseByStatus']);
 
@@ -50,19 +48,12 @@ Route::group(['prefix' => 'admin'], function () {
     Voyager::routes();
 });
 
-//Moodle Requests
+//Requests
 Route::controller(SupportTicketController::class)->group(function () {
     //Creating new request (for student student)
     Route::get('/SupportTicketNew', 'sendNew');
     Route::post('/SupportTicketNew', 'insert');
 });
-
-//Dormitory Requests
-// Route::controller(DormitoryTicket::class)->group(function () {
-//     //Creating new request (for student student)
-//     Route::get('/DormitoryTicketNew', 'sendNew');
-//     Route::post('/DormitoryTicketNew', 'inProgress');
-// });
 
 Route::get('/SupportTicketStatus', [SupportTicketController::class, 'supportTicketStatusCookie']);
 Route::post('/SupportTicketStatus', [SupportTicketController::class, 'supportTicketStatus']);